# apps/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: PhyloPic API
  description: Documentation for the next version of *PhyloPic*'s Application Programming Interface.
  contact:
    name: Mike Keesey
    email: keesey+phylopic@gmail.com
    url: http://tmkeesey.net
  license:
    name: CC0
    url: https://creativecommons.org/share-your-work/public-domain/cc0
  version: 0.0.0
...
servers:
  - url: https://api.phylopic.org
    description: The production API
components:
  headers:
...
    Access-Control-Allow-Credentials:
      description: Specifies that credentials are allowed in requests. Always `true`.
      example: "true"
      schema:
        pattern: "^true$"
        type: string
    Access-Control-Allow-Origin:
      description: Specifies that requests may be made from any origin. Always `*`.
      example: "*"
      schema:
        pattern: '^\*$'
        type: string
    Cache-Control:
      description: Specifications for caching mechanisms.
      example: no-cache
      schema:
        type: string
    Content-Length:
      description: The length of the content, in bytes.
      example: 1024
      schema:
        format: number
        type: string
...
    Content-Type:
      description: The media type for the payload, including API version. Always `application/vnd.phylopic.v2+json`.
      example: application/vnd.phylopic.v2+json
      schema:
        format: mimetype
        type: string
    ETag-GET:
      description: An identifier for this version of the resource. May be used on subsequent calls to avoid unnecessary reloading (`If-None-Match` header).
      example: '"identifier"'
      schema:
        pattern: '^"[^"]+"$'
        type: string
...
    Location:
      description: The URL for the resource.
      schema:
        format: url
        type: string
...
  parameters:
    Accept:
      name: Accept
      in: header
      description: |
        The media type for the response, including API version. Must allow `application/vnd.phylopic.v2+json`.

        If omitted, `*/*` (a wildcard allowing all types) is assumed, and the response's `Content-Type` header will be `application/vnd.phylopic.v2+json`.
      schema:
        format: mimetype
        type: string
...
    If-Match-GET:
      name: If-Match
      in: header
      description: Identifies which version(s) of the list is/are expected, as given by the `ETag` header on a previous response. If that version is out of date, a `416` response will be returned.
      schema:
        pattern: '^(\*|(W\\/)?"([^"]|\\")+"(,\s*(W\/)?"([^"]|\\")+")*)$'
        type: string
      example: '"identifier"'
...
    If-None-Match:
      name: If-None-Match
      in: header
      description: Identifies which version of the resource has already been loaded, as given by the `ETag` header on a previous response. If this matches the current version, a `304` response will be returned.
      schema:
        pattern: '^(\*|(W\\/)?"([^"]|\\")+"(,\s*(W\/)?"([^"]|\\")+")*)$'
        type: string
    Range:
      name: Range
      in: header
      required: true
      description: |
        A range of the items to read, inclusive.

        Example: `items=0-9`
      schema:
        pattern: '^items=\d+-\d+$'
        type: string
    contributor:
      name: contributor
      in: query
      required: false
      description: |
        Limits results to entries contributed by a certain user.

        The value is the contributor's email address.
      schema:
        format: email
        type: string
    created_lt:
      name: created_lt
      in: query
      required: false
      description: |
        Limits results to entries created at or before this time.

        The value must be an ISO-formatted date-time, for example, `1970-01-01T00:00:00.000Z`.
      schema:
        format: date-time
        type: string
    created_gt:
      name: created_gt
      in: query
      required: false
      description: |
        Limits results to entries created at or after this time.

        The value must be an ISO-formatted date-time, for example, `1970-01-01T00:00:00.000Z`.
      schema:
        format: date-time
        type: string
    embed-Image:
      name: embed
      in: query
      description: |
        A space-separated list of related entities to embed. May include:
        - `generalNode`
        - `nodes`
        - `specificNode`

        If omitted, no entities will be embedded.
      schema:
        pattern: "^((generalNode|nodes|specificNode) )*(generalNode|nodes|specificNode)$"
        type: string
...
    licensecomponents:
      name: licensecomponents
      in: query
      required: false
      description: |
        Limits entries to those with or without certain license components.

        Components are represented as a space-separated list. May include:
        - `by` (Attribution)
        - `nc` (NonCommercial)
        - `sa` (StandAlone)

        Any of these may be prepended by `-` to indicate that only images under licenses **without** that component should be returned.

        Examples:
        - `by`: only include images with licenses that require attribution.
        - `-by`: exclude all images with licenses that require attribution.
        - `-by -nc -sa`: only include images in the public domain (having none of the license components).
      schema:
        format: '^(\-?(by|nc|sa) )*(\-?(by|nc|sa))$'
        type: string
    modified_lt:
      name: modified_lt
      in: query
      required: false
      description: |
        Limits results to entries last modified at or before this time.

        The value must be an ISO-formatted date-time, for example, `1970-01-01T00:00:00.000Z`.
      schema:
        format: date-time
        type: string
      example: "1970-01-01T00:00:00.000Z"
    modified_gt:
      name: modified_gt
      in: query
      required: false
      description: |
        Limits results to entries last modified at or after this time.

        The value must be an ISO-formatted date-time, for example, `1970-01-01T00:00:00.000Z`.
      schema:
        format: date-time
        type: string
    sort:
      name: sort
      in: query
      required: false
      description: |
        A space-separated list of fields to sort by. May include:
        - `created`
        - `modified`

        Either of these may be prepended by `-` to indicate descending order. The default value is `-modified`.
      schema:
        pattern: '^(\-?(created|modified) )*(\-?(created|modified))$'
        type: string
...
    uuid-Entity:
      name: uuid
      in: path
      required: true
      description: The UUID of the entity.
      schema:
        format: uuid
        type: string
  responses:
...
    "304":
      description: The resource has not been modified since the last time it was requested.
      headers:
        Access-Control-Allow-Credentials:
          $ref: "#/components/headers/Access-Control-Allow-Credentials"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        ETag:
          $ref: "#/components/headers/ETag-GET"
...
    "400":
      description: There is an error in the request.
      headers:
        Access-Control-Allow-Credentials:
          $ref: "#/components/headers/Access-Control-Allow-Credentials"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        Content-Length:
          $ref: "#/components/headers/Content-Length"
        Content-Type:
          $ref: "#/components/headers/Content-Type"
      content:
        application/vnd.phylopic.v2+json:
          schema:
            $ref: "#/components/schemas/Errors"
    "404":
      description: The resource cannot be found.
      headers:
        Access-Control-Allow-Credentials:
          $ref: "#/components/headers/Access-Control-Allow-Credentials"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        Cache-Control:
          $ref: "#/components/headers/Cache-Control"
        Content-Length:
          $ref: "#/components/headers/Content-Length"
        Content-Type:
          $ref: "#/components/headers/Content-Type"
      content:
        application/vnd.phylopic.v2+json:
          schema:
            $ref: "#/components/schemas/Errors"
    "406":
      description: The request cannot be accepted because the `Accept` header does not allow an available type of content.
      headers:
        Access-Control-Allow-Credentials:
          $ref: "#/components/headers/Access-Control-Allow-Credentials"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        Content-Length:
          $ref: "#/components/headers/Content-Length"
        Content-Type:
          $ref: "#/components/headers/Content-Type"
      content:
        application/vnd.phylopic.v2+json:
          schema:
            $ref: "#/components/schemas/Errors"
    "410":
      description: The resource has been removed.
      headers:
        Access-Control-Allow-Credentials:
          $ref: "#/components/headers/Access-Control-Allow-Credentials"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        Cache-Control:
          $ref: "#/components/headers/Cache-Control"
        Content-Length:
          $ref: "#/components/headers/Content-Length"
        Content-Type:
          $ref: "#/components/headers/Content-Type"
      content:
        application/vnd.phylopic.v2+json:
          schema:
            $ref: "#/components/schemas/Errors"
...
    "416":
      description: The requested range is not satisfiable.
      headers:
        Access-Control-Allow-Credentials:
          $ref: "#/components/headers/Access-Control-Allow-Credentials"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        Content-Length:
          $ref: "#/components/headers/Content-Length"
        Content-Type:
          $ref: "#/components/headers/Content-Type"
        ETag:
          $ref: "#/components/headers/ETag-GET-Range"
      content:
        application/vnd.phylopic.v2+json:
          schema:
            $ref: "#/components/schemas/Errors"
    "list-Image":
      description: |
        [HAL](http://stateless.co/hal_specification.html) [JSON](https://www.json.org/) list of embedded image records.
      headers:
        Accept-Ranges:
          $ref: "#/components/headers/Accept-Ranges"
        Access-Control-Allow-Credentials:
          $ref: "#/components/headers/Access-Control-Allow-Credentials"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        Cache-Control:
          $ref: "#/components/headers/Cache-Control"
        Content-Length:
          $ref: "#/components/headers/Content-Length"
        Content-Range:
          $ref: "#/components/headers/Content-Range"
        Content-Type:
          $ref: "#/components/headers/Content-Type"
        ETag:
          $ref: "#/components/headers/ETag-GET-Range"
        Warning:
          $ref: "#/components/headers/Warning-GET-Range"
      content:
        application/vnd.phylopic.v2+json:
          schema:
            type: object
            required:
              - _embedded
              - _links
            properties:
              _embedded:
                type: object
                description: Container for embedded entities.
                properties:
                  items:
                    type: array
                    description: The requested image entities.
                    items:
                      $ref: "#/components/schemas/ImageWithEmbedded"
              _links:
                type: object
                description: Container for a link to this resource (`self`).
                required:
                  - self
                properties:
                  self:
                    allOf:
                      - $ref: "#/components/schemas/Link"
                      - description: Link to this list.
              total:
                description: The total number of items in the set.
                format: uint
                type: number
...
  schemas:
...
    ImageWithEmbedded:
      description: Resource for an image.
      required:
        - _embedded
      allOf:
        - type: object
          properties:
            _embedded:
              type: object
              description: Embedded entities related to the image. See `_links` for more information.
              properties:
                generalNode:
                  allOf:
                    - $ref: "#/components/schemas/Node"
                    - nullable: true
                nodes:
                  type: array
                  items:
                    $ref: "#/components/schemas/Node"
                specificNode:
                  $ref: "#/components/schemas/Node"
        - $ref: "#/components/schemas/Image"
...
    TitledLink:
      description: A link with a title, usually for an external resource.
      allOf:
        - $ref: "#/components/schemas/Link"
        - type: object
          required:
            - title
          properties:
            title:
              type: string
              example: "Name"
  securitySchemes:
    keyAuth:
      description: User consuming the API data.
      type: apiKey
      in: header
      name: X-API-Key
paths:
...
  /images:
    get:
      operationId: getImages
      summary: Lists all images.
      tags:
        - images
      parameters:
        - $ref: "#/components/parameters/Accept"
        - $ref: "#/components/parameters/If-Match-GET"
        - $ref: "#/components/parameters/Range"
        - $ref: "#/components/parameters/contributor"
        - $ref: "#/components/parameters/created_lt"
        - $ref: "#/components/parameters/created_gt"
        - $ref: "#/components/parameters/embed-Image"
        - $ref: "#/components/parameters/licensecomponents"
        - $ref: "#/components/parameters/modified_lt"
        - $ref: "#/components/parameters/modified_gt"
        - $ref: "#/components/parameters/sort"
      security:
        - keyAuth: []
      responses:
        "206":
          $ref: "#/components/responses/list-Image"
        "400":
          $ref: "#/components/responses/400"
        "406":
          $ref: "#/components/responses/406"
        "416":
          $ref: "#/components/responses/416"
  /images/{uuid}:
    get:
      operationId: getImage
      summary: Returns data for a specific image.
      tags:
        - images
      parameters:
        - $ref: "#/components/parameters/Accept"
        - $ref: "#/components/parameters/If-None-Match"
        - $ref: "#/components/parameters/uuid-Entity"
        - $ref: "#/components/parameters/embed-Image"
      security:
        - keyAuth: []
      responses:
        "200":
          description: |
            [HAL](http://stateless.co/hal_specification.html) [JSON](https://www.json.org/) image record.
          headers:
            Access-Control-Allow-Credentials:
              $ref: "#/components/headers/Access-Control-Allow-Credentials"
            Access-Control-Allow-Origin:
              $ref: "#/components/headers/Access-Control-Allow-Origin"
            Cache-Control:
              $ref: "#/components/headers/Cache-Control"
            Content-Length:
              $ref: "#/components/headers/Content-Length"
            Content-Type:
              $ref: "#/components/headers/Content-Type"
            ETag:
              $ref: "#/components/headers/ETag-GET"
          content:
            application/vnd.phylopic.v2+json:
              schema:
                $ref: "#/components/schemas/ImageWithEmbedded"
        "304":
          $ref: "#/components/responses/304"
        "404":
          $ref: "#/components/responses/404"
        "406":
          $ref: "#/components/responses/406"
        "410":
          $ref: "#/components/responses/410"
  /licenses:
    get:
      operationId: getLicenses
      summary: Returns a list of licenses.
      tags:
        - images
      description:
        Returns a list of all licenses that may be used for new submissions.

        Note that this does not include all licenses for all images in the database.
      parameters:
        - $ref: "#/components/parameters/Accept"
      security:
        - keyAuth: []
      responses:
        "200":
          description: List of titled links to licenses.
          headers:
            Access-Control-Allow-Credentials:
              $ref: "#/components/headers/Access-Control-Allow-Credentials"
            Access-Control-Allow-Origin:
              $ref: "#/components/headers/Access-Control-Allow-Origin"
            Cache-Control:
              $ref: "#/components/headers/Cache-Control"
            Content-Length:
              $ref: "#/components/headers/Content-Length"
            Content-Type:
              $ref: "#/components/headers/Content-Type"
            ETag:
              $ref: "#/components/headers/ETag-GET"
          content:
            application/vnd.phylopic.v2+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TitledLink"
        "406":
          $ref: "#/components/responses/406"
  /lineage/{uuid}:
    get:
      operationId: getLineage
      summary: Retrieves the ancestral lineage for a given node.
      description: The nodes are listed in order from least to most ancestral. That is, the first node is the specified node, the second is its parent node, and the last is the root node.
      tags:
        - nodes
      parameters:
        - $ref: "#/components/parameters/Accept"
        - $ref: "#/components/parameters/If-Match-GET"
        - $ref: "#/components/parameters/Range"
        - $ref: "#/components/parameters/embed-Node"
        - $ref: "#/components/parameters/uuid-Entity"
      security:
        - keyAuth: []
      responses:
        "206":
          $ref: "#/components/responses/list-Node"
        "400":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/404"
        "406":
          $ref: "#/components/responses/406"
        "410":
          $ref: "#/components/responses/410"
        "416":
          $ref: "#/components/responses/416"
...
  /resolve:
    get:
      operationId: resolve
      summary: Resolves an external URI to the closest phylogenetic node.
      tags:
        - nodes
        - search
      parameters:
        - $ref: "#/components/parameters/Accept"
        - $ref: "#/components/parameters/embed-Node"
        - name: uri
          in: query
          required: true
          description: The URI of the external resource.
          schema:
            format: uri
            type: string
          example: https://eol.org/pages/328046
      security:
        - keyAuth: []
      responses:
        "307":
          description: A corresponding or ancestral node was found.
          headers:
            Location:
              $ref: "#/components/headers/Location"
        "400":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/404"
        "406":
          $ref: "#/components/responses/406"
...